♻️ refactor: metadata를 위한 서버컴포넌트 변환 및 로고 수정, 이메일 인증 콘솔 제거, 파비콘 생성, 로그인인풋 디폴트값 생성
